LPS-126038 Refactor code for symmetry with DatasetViewContentRendererNames
This makes the API easier to navigate, understand and document since we have
the list of all possible built-in content renderer names in one single place.
